minor setup changes

+++ Backend/clearDB.py
@@ -1,7 +1,9 @@
 import pymongo
 
 client = pymongo.MongoClient("mongodb://mongo:27017/")
-allDBs = client.list_database_names()
 
-if 'convrt_database' in allDBs:
-    client.drop_database('convrt_database')
+if client:
+    allDBs = client.list_database_names()
+
+    if 'convrt_database' in allDBs:
+        client.drop_database('convrt_database')
